Compute and visualize energy for the selected frequency range.
Use p5.FFT to compute the energy contained in the selected frequency
band. Currently this is being displayed in a rolloing time-series
scatter plot that displays the most recent 5 seconds of data.

An "activation threshold" can also be configured, with the idea being
that energies that exceed the threshold should be registered as a
"percussion hit".